Kumar dev

+++ app/Http/Controllers/Admin/ClientsController.php
@@ -8,6 +8,8 @@ use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Mail;
+use App\Mail\SendLoginDetail;
 use App\Models\Customer;
 use App\Models\Device;
 use App\Models\CustomerDevice;
@@ -65,6 +67,17 @@ class ClientsController extends Controller
                         'role' => 'user',
                     ]);
 
+            // Prepare email data
+            $loginData = [
+                'title' => 'Your Login Details Have Been Create',
+                'body' => 'your login details.',
+                'email' => $user->email,
+                'password' => $request->password,
+            ];
+
+            // Send email
+            Mail::to($user->email)->send(new SendLoginDetail($loginData));
+
 
             // Create a new Customer
             $customer = new Customer;
@@ -93,7 +106,7 @@ class ClientsController extends Controller
             DB::commit();
     
             // Redirect with success message
-            return redirect()->route('clients')->with('success', 'Client added successfully.');
+            return redirect()->route('clients')->with(['success' => 'Client added successfully.']);
         } catch (\Exception $e) {
             /*-------- Rollback Database Entry --------*/
             DB::rollback();
@@ -216,12 +229,11 @@ class ClientsController extends Controller
 
     public function destroy($id)
     {
-        try {
-            $customer = Customer::findOrFail($id);
+         $customer = Customer::findOrFail($id);
+        if($customer){
             $customer->delete();
-    
-            return redirect()->route('clients')->with('success', 'Customer deleted successfully.');
-        } catch (\Exception $e) {
+            return redirect()->route('clients')->with(['success' => 'Customer deleted successfully.']);
+        }else{
             return back()->withErrors(['error' => 'An error occurred while deleting the customer.']);
         }
     }
